+++ include/test.h
@@ -2,7 +2,7 @@
 #define __TEST_H__
 
 void test_randvar_uniform();
-void test_randvar_exp(float lambda);
+void test_randvar_exp(double lambda);
 void test_es_pq();
 
 #endif // __TEST_H__
